refactor: move data from student to admin
